Refactor schematic type declarations into dedicated file
